Fix type returned by FillWithSameByte.toUint64
Before these changes the code was still compiling, automatic type conversions appear to be easy to miss.

+++ test/FillWithSameByte.sol
@@ -14,8 +14,8 @@ library FillWithSameByte {
         return uint32(repeatByte(b, 4));
     }
 
-    function toUint64(uint8 b) public pure returns (uint32) {
-        return uint32(repeatByte(b, 8));
+    function toUint64(uint8 b) public pure returns (uint64) {
+        return uint64(repeatByte(b, 8));
     }
 
     function toUint256(uint8 b) public pure returns (uint256) {
